Replace more serious bug with less serious bug
After reading the "What type of bugs are allowed?" topic on Piazza,
I decided to replace a more serious bug (which caused a segfault when
playdom was executed) with a less serious one. The new bug allows for
playdom to still play through an entire game without problems (which
seems more appropriate with regard to the instructor's commentary).