Update notes on Adventurer bug based on testing
After having tested this on the Flip servers, it appears that the
test scenario that causes a crash may be somewhat platform-dependent.
That is, the test scenario will crash on Mac OS X (10.11) but it
won't on the Flip servers (Redhat Linux).  I've set up the test so
that it will properly run in the parent process if it hasn't crashed
when executed in a child process, so it should be okay.