Fix handCount tests in cardtest1 (Adventurer)
The cardtest1 tests were originally designed around the idea that
Adventurer would not be discarding the played card, so the handCount
checks did not subtract one from the expected value to account for
this.  This allowed me to check for other behavior *despite* this
bug in Adventurer, so it was appropriate then.  I'm planning on
fixing this bug in the Adventurer code, so this is essentially prep
work for that forthcoming change.