Add handPos generation to randomtestadventurer.c
This is also related to a forthcoming bug fix that will make the
code for the Adventurer card discard the played card.

The random tester was previously adding the Adventurer card in a
random (valid) position in the hand but it wasn't keeping track of
the handPos.  We'll need the handPos in the future, so this makes
things a little easier for when that time comes.