Improve Mine's choice2 card validation
The code for the Mine card ensures that choice2 (the card to upgrade
to) is a valid card (i.e. within the range of values used by
dominion.h's CARD enum) but it doesn't do anything to ensure that
choice2 is a treasure card.  As such, you could potentially upgrade
a treasure card to a non-treasure card, which goes against Dominion's
rules.

This fix brings the choice2 validation in line with the choice1
validation and fixes this bug.