Modify Mine test's trash/discard check
The previous tests to check whether Mine only affects one instance of
the card to upgrade (choice1) didn't always work as planned.  If the
bug that causes Mine to discard rather than trash the choice1 card is
fixed then the check to only trash one card would erroneously fail.
This resolves the issue but I'm not quite sure that I'm sold on this
method.